Upload file to mongodb

+++ api/controllers/test/AnswerController.js
@@ -22,15 +22,20 @@ module.exports = {
     delete: function (req, res) {
         var id = req.param('id');
         Answer.find({idQuestion: id}).exec(function createCB(err, answers) {
-            answers.forEach(function (item, index) {
-                Answer.destroy({id: item.id}).exec(function createCB(err, result) {
-                    if (err) sails.log(err)
-                    else {
-                        if (index == answers.length - 1)
-                            res.send('Answers of message have been deleted too!');
-                    }
+            if (answers.length == 0) {
+                res.send('Question no has answers delete!')
+            }
+            else {
+                answers.forEach(function (item, index) {
+                    Answer.destroy({id: item.id}).exec(function createCB(err, result) {
+                        if (err) sails.log(err);
+                        else {
+                            if (index == answers.length - 1)
+                                res.send('Answers of message have been deleted too!');
+                        }
+                    })
                 })
-            })
+            }
         });
     }
 };
